fix: encode uses shared ref, update default trait derives (#11)
Changes the `encode()` function to use a shared reference, rather than
a mutable one, as it only needs to modify the passed PDU data.

Adds `Copy` and `Clone` to the default trait derivations, as the
generated structs are trivially copyable.

Fixes typo in README.

Updates changelog and bumps crate version.